requirements.txt
huggingface-hub[cli]==0.25.2
openai==1.51.2
python-dotenv==1.0.1
transformers==4.45.2
loguru==0.7.2
ray[default]==2.37.0
tqdm==4.66.5
pandas==2.2.3
easydict==1.13
backoff==2.2.1
httpx[socks]==0.27.2
rank-bm25==0.2.2
hydra-core>=1.0.0
sexpdata==1.0.0
pylspclient==0.0.3
pampy==0.3.0
jsonlines==4.0.0
dataclasses-json==0.6.7
parglare==0.18.0
protobuf==3.20.1
tiktoken==0.7.0
accelerate==1.0.1
blobfile==3.0.0
sentencepiece==0.2.0
